+++ src/components/header/breadcrumbs/BreadcrumbsObj.ts
@@ -1,0 +1,6 @@
+export const breadcrumbsObj: Record<string, string> = {
+    '/': 'Главная',
+    juiciest: 'Самое сочное',
+    'vegan-cuisine': 'Веганская кухня',
+    '/vegan-cuisine/main-courses': 'Вторые блюда',
+};
